ap {HydroData}	R Documentation

Airports

Description

Dataset containing information global Airports subset to the Daymet Domain

Usage

ap

Format

a dataframe instance, 1 row per station with columns:

'name': A character Name of airport. May or may not contain the City name.

'city': A character Main city served by airport. May be spelled differently from Name.

'country': A character Country or territory where airport is located.

'IATA': A character 3-letter IATA code

'ICAO': A numeric 4-letter ICAO code

'lat': A numeric Latitude of airport

'lon': A numeric Longitude of airport

Source

OpenFlights

Examples

## Not run: 
 airports = HydroData::ap

## End(Not run)

[Package HydroData version 0.0.7 Index]