findAirports {HydroData}	R Documentation

Find Airport Data

Description

findAirports returns a SpatialPointsDataframe of all airports within an AOI. Data comes from the Openflights database and includes the following attributes:

'name' : character Name of airport. May or may not contain the City name.

'city' : character Main city served by airport. May be spelled differently from Name.

'country': character Country or territory where airport is located.

'IATA' : character 3-letter IATA code

'ICAO' : numeric 4-letter ICAO code

'lat' : numeric Latitude of airport

'lon' : numeric Longitude of airport

Usage

findAirports(AOI = NULL, ids = FALSE)

Arguments

AOI
A Spatial* or simple features geometry, can be piped from getAOI

ids
If TRUE, a vector of airport ICAO codes is added to retuned list (default = FALSE)

Value

a list of minimum length 2: AOI and ap

Author(s)

Mike Johnson

Examples

## Not run: 
ap = getAOI(state = "CO", county = "El Paso") %>% findAirports()

## End(Not run)

[Package HydroData version 0.0.7 Index]