findCDL {HydroData}	R Documentation

Find Cropland Land Cover Data from the Cropand Data Layer

Description

findCLD returns Raster land cover data from the USDA Cropland Data Layer Dataset (Cropscape) for an AOI.

Usage

findCDL(AOI, year = 2017)

Arguments

AOI
A Spatial* or simple features geometry, can be piped from getAOI

year
the year(s) to download. Options include 2008-2017 for most locations. Default = 2017

Value

a list() of minimum length 2: AOI and cld

Author(s)

Mike Johnson

Examples

## Not run: 
 dt = getAOI(clip = list("Devil Tower")) %>% findCDL()
 dt = getAOI(clip = list("Devil Tower")) %>% findCDL(2011:2016)

## End(Not run)

[Package HydroData version 0.0.7 Index]