findGAGESII {HydroData}	R Documentation

Find GAGESII points and basins

Description

The 'Geospatial Attributes of Gages for Evaluating Streamflow, version II' (GAGESII) provides geospatial data and classifications for 9,322 stream gages maintained by the U.S. Geological Survey (USGS). findGAGESII returns a SpatialPointsDataFrame* of GAGESII outlets within an AOI. Data comes from the USGS CIDA server and contains the following attributes:

'id' : character Internal feature number

'STAID' : character USGS NWIS Station ID

'STANAME': character USGS NWIS Station Name

'CLASS' : character Classification (Ref or Non-ref)

'AGGECOREGION' : numeric Aggregated ecoregion

'DRAIN_SQKM' : numeric Drainage area, sq km

'HUC02' : numeric Hydrologic Unit Code, 2-digit

'LAT_GAGE' : character

'LNG_GAGE' : character Longitude, decimal degrees

'STATE': character State at gage location

'HCDN_2009' : character If gage is part of HCDN-2009

'ACTIVE09' : numeric If gage active in water year 2009

'FLYRS1990' : numeric Number of complete years of flow data between 1900 and 2009

'FLYRS1950' : numeric Number of complete years of flow data between 1950 and 2009

'FLYRS1990' : numeric Number of complete years of flow data between 1990 and 2009

If basins = TRUE a SpatialPolygonsDataFrame of the gage drainage basin will also be appended to the returned list and contain the following attributes:

'id' : character Internal feature number and data identifier

'ogr_fid' : character Internal feature number

'area': character Basin Area

'perimeter' : character Basein Parameters

'gage_id' : numeric USGS NWIS Station ID

Usage

findGAGESII(AOI = NULL, basins = FALSE, ids = FALSE)

Arguments

AOI
A Spatial* or simple features geometry, can be piped from getAOI

basins
If TRUE, returns a list of GAGESII basin in addition

ids
If TRUE, a vector of gage IDs is added to retuned list (default = FALSE)

Value

a list() of minimum length 2: AOI and gagesII

Author(s)

Mike Johnson

Examples

## Not run: 
#Get GAGESII outlets for AOI
bas = getAOI(clip = list("UCSB", 10, 10)) %>%  findGAGESII()

#Get GAGESII outlets and basins for AOI
bas = getAOI(clip = list("UCSB", 10, 10)) %>%  findGAGESII(basins = TRUE)

## End(Not run)

[Package HydroData version 0.0.7 Index]