Find NOAA Maintained GHCN Stations

findGHCN returns a SpatailPointsDataframe of all Global Historical Climatology Network (GHCN) stations within an AOI. Data comes from the NOAA and includes the following attributes:

'ID' : character Station ID

'NAME' : character Station Name

'LAT': numeric Latitude, decimal degrees

'LON' : numeric Longitude, decimal degrees

'PARAMETER' : character Observed parameter

'START_YEAR' : integer Year of first observation

'END_YEAR' : integer Year of last observation

findGHCN(AOI = NULL, param = NULL, ids = FALSE)

Arguments

AOI
A Spatial* or simple features geometry, can be piped from getAOI

param
Returned results can be filtered by a requested parameter valid parameters can be found calling unique(ghcn_stations$PARAMETER)

ids
If TRUE, a vector of station IDs codes is added to retuned list (default = FALSE)

Value

a list() of minimum length 2: AOI and a

Examples

# NOT RUN {
# Get all stations and parameters
sta = getAOI(state = "CO", county = "El Paso") %>% findGHCN()

#Limit call to precipiation gages
prcp.sta = getAOI(state = "CO", county = "El Paso") %>% findGHCN(param = 'PRCP')

# }