findNED {HydroData}	R Documentation

Find National Elevation Data (NED)

Description

findNED returns Raster elevation data from the National Elevation Dataset (NED) for an AOI. Data comes the the USA National Map.

Usage

findNED(AOI = NULL, res = 1)

Arguments

AOI
A Spatial* or simple features geometry, can be piped from getAOI

res
resolution of NED data. 1 equals 1 arc second, 13 equals 1/3 arc second.

Value

a list() of minimum length 2: AOI and NED

Author(s)

Mike Johnson

Examples

## Not run: 
 el.paso.elev = getAOI(state = "CO", county = "El Paso") %>% findNED(res = 1)

## End(Not run)

[Package HydroData version 0.0.7 Index]