findNHD {HydroData}	R Documentation

Find National Hydrography River Networks

Description

findNHD returns a SpatialLinesDataframe of all NHDFlowlines reaches within an AOI. Data comes from the USGS CIDA server and contain 90 attributes, perhaps most notably:

'comid' : integer Integer value that uniquely identifies the occurrence of each feature in the NHD

'reachcode' : character Unique identifier for a ‘reach’. The first eight numbers are the WBD_HUC8

'fdate': POSITct Date of last feature modification

'gnis_id' : character Unique identifier assigned by GNIS

'gnis_name' : character Proper name, term, or expression by which a particular geographic entity is known

'lengthkm' : numeric Length of linear feature based on Albers Equal Area

'areasqkm' : numeric Area of areal feature based on Albers Equal Area,

'flowdir' : character Direction of flow relative to coordinate order.

'wbareacomi' : integer The COMID of the waterbody through which the flowline flows.

'ftype': character unique identifier of a feature type

Usage

findNHD(AOI = NULL, comid = NULL, nwis = NULL, streamorder = NULL,
  crop = T)

Arguments

AOI
A Spatial* or simple features geometry, can be piped from getAOI

ids
If TRUE, a vector of NHD COMIDs is added to retuned list (default = FALSE)

Value

a list() of minimum length 2: AOI and nhd

Author(s)

Mike Johnson

Examples

## Not run: 
nhd  = getAOI(clip = list("Tuscaloosa, AL", 10, 10)) %>% findNHD()

## End(Not run)

[Package HydroData version 0.0.7 Index]