findNID {HydroData}	R Documentation

Find Dams in the US Army Core National Inventory (NID)

Description

findNID returns a SpatialPointsDataFrame of all US Army Corps Dams for an Area of Interest from the National Inventory of Dams dataset. The National Inventory of Dams (NID) is a congressionally authorized database documenting dams in the United States and its territories This dataset is accessed through the dams R package and contains 61 attributes, perhaps most notably:

'Dam_Name' : character Dam Name

'NID_ID' : character Unique ID for the dam

'River': character Name of the river

'Owner_Type' : character Type of Owner

'Dam_Type' : character Type of Dam

'Primary_Purpose' : numeric Primary Purpose served

'Dam_Length' : numeric Length of the dam

'Dam_Height' : numeric Height of the dam

'Max_Discharge' : numeric Maximum Discharge

'Max_Storage': character Maximum Storage

'Normal_Storage': character Normal Storage

Usage

findNID(AOI = NULL, ids = FALSE)

Arguments

AOI
A Spatial* or simple features geometry, can be piped from getAOI

ids
If TRUE, a vector of Dam IDs is added to retuned list (default = FALSE)

Value

a list() of minimum length 2: AOI and dams

Author(s)

Mike Johnson

Examples

## Not run: 
# Find all dams in Texas
tx.dams = getAOI(state = "TX") %>% findNID()

## End(Not run)

[Package HydroData version 0.0.7 Index]