findNLCD {HydroData}	R Documentation

Find National Land Cover Products (NLCD)

Description

findNLCD returns Raster land cover data from the National Land Cover Dataset (NLCD) for an AOI. Data comes the the USA National Map and is avaialble for years 2001, 2006, 2011. In additon to landcover, users can get data reflecting impervious surface and conaopy cover.

Usage

findNLCD(AOI = NULL, year = 2011, type = "landcover")

Arguments

AOI
A Spatial* or simple features geometry, can be piped from getAOI

year
the year(s) to download. Options include 2001, 2006, 2011. Default = 2011

type
the type of data to downlaod. Options include landcover, canopy, and impervious. Default = landcover

Value

a list() of minimum length 2: AOI and NLCD

Author(s)

Mike Johnson

Examples

## Not run: 
 dt = getAOI(clip = list("Devil Tower")) %>% findNLDC(2006, 'landcover')
 dt = getAOI(clip = list("Devil Tower")) %>% findNLDC(2011, 'canopy')
 dt = getAOI(clip = list("Devil Tower")) %>% findNLDC(2011, 'impervious')

## End(Not run)

[Package HydroData version 0.0.7 Index]