findNLDI {HydroData}	R Documentation

Get NLDI Data

Description

This function acts as a formal query to the NLDI

Usage

findNLDI(comid = NULL, nwis = NULL, huc12 = NULL, type = NULL,
  spatial = TRUE, resources = NULL)

Arguments

comid
a COMID value (see findNHD and findNearestCOMID)

nwis
a USGS station values (see findUSGS and findNearestUSGS)

huc12
a HUC12 id (see findHUC)

type
a type of data to return ("UT", "UM","DD", "DM", "basin")

spatial
if TRUE returned data will be sp, else of type sf

resources
what type of resources to add ("comid", "huc12pp", "npdes_rad", "nwis", "wqp")

Value

a list of spatial / sf features

Author(s)

Mike Johnson

[Package HydroData version 0.0.7 Index]