findNearestCOMID {HydroData}	R Documentation

Find Neareast COMIDs

Description

findNearestCOMID returns a SpatialPointsDataFrame of the 'n' number of COMIDs closest to a declared point.

Usage

findNearestCOMID(point = NULL, n = 5, ids = FALSE,
  streamorder = NULL, bb = FALSE)

Arguments

point
a point described by lat/long, can be piped from geocode

n
the number of COMIDs to find (default = 5)

ids
If TRUE, a vector of NHD COMIDs is added to retuned list (default = FALSE)

bb
If TRUE, the geometry of the minimum bounding area of the features is added to returned list (default = FALSE)

Value

a list() of minimum length 2: AOI and ap

Author(s)

Mike Johnson

See Also

findAirports

Examples

## Not run: 
pt = geocode("University of Oregon") %>% findNearestCOMID(n = 5)

## End(Not run)

[Package HydroData version 0.0.7 Index]