Find Neareast GHCN stations

findNearestGHCN returns a SpatialPointsDataFrame of the 'n' number of GHCN station closest to a declared point.

findNearestGHCN(point = NULL, n = 5, parameters = NULL, ids = FALSE,
  bb = FALSE)

Arguments

point
a point described by lat/long, can be piped from geocode

n
the number of stations to find (default = 5)

parameters
Can be used to filter the returned stations by measurement types

ids
If TRUE, a vector of airport stations IDs is added to retuned list (default = FALSE)

bb
If TRUE, the geometry of the minimum bounding area of the features is added to returned list (default = FALSE)

Value

a list() of minimum length 2: AOI and ghcn

See also

Examples

# NOT RUN {
pt = geocode("UCSB") %>% findNearestGHCN(n = 5)
# }