findNearestHUC12 {HydroData}	R Documentation

Find Neareast HUC12 boundaries

Description

findNearestHUC12 returns a SpatialPolygonDataFrame of the 'n' number of HUC12 basins closest to a declared point.

Usage

findNearestHUC12(point = NULL, n = 1, ids = FALSE, bb = FALSE)

Arguments

point
a point described by lat/long, can be piped from geocode

n
the number of basins to find (default = 5)

ids
If TRUE, a vector of basin HUC IDs is added to retuned list (default = FALSE)

bb
If TRUE, the geometry of the minimum bounding area of the features is added to returned list (default = FALSE)

Value

a list() of minimum length 2: AOI and wbd

Author(s)

Mike Johnson

See Also

findWBD

Examples

## Not run: 
pt = geocode("UCSB") %>% findNearestHUC12(n = 5)

## End(Not run)

[Package HydroData version 0.0.7 Index]