findNearestNWIS {HydroData}	R Documentation

Find Neareast NWIS stations

Description

findNearestNWIS returns a SpatialPointsDataFrame of the 'n' number of NWIS gages closest to a declared point.

Usage

findNearestNWIS(point = NULL, n = 5, ids = FALSE, bb = FALSE, ...)

Arguments

point
a point described by lat/long, can be piped from geocode

n
the number of NWIS gages to find (default = 5)

ids
If TRUE, a vector of station IDs is added to retuned list (default = FALSE)

bb
If TRUE, the geometry of the minimum bounding area of the features is added to returned list (default = FALSE)

Value

a list() of minimum length 2: AOI and nwis

Author(s)

Mike Johnson

See Also

findNWIS

Examples

## Not run: 
pt = geocode("UCSB") %>% findNearestNWIS(n = 5)

## End(Not run)

[Package HydroData version 0.0.7 Index]