findRoads {HydroData}	R Documentation

Find US Census Bureau TIGER Road Networks

Description

findRoads returns a SpatialLinesDataFrame of TIGER road networks cropped to an Area of Interest. This dataset is accessed through the US Census geo portal and contains the following attributes:

'LINEARID' : character Linear feature identifier

'FULLNAME' : character Feature Name

'RTTYP': character Route Type Code

'MTFCC' : character Road Classification

S1100 Primary Roads

R1011 Railroad Feature (Main, Spur, or Yard)

R1051 Carline, Streetcar Track, Monorail, Other Mass Transit Rail)

R1052 Cog Rail Line, Incline Rail Line, Tram

S1100 Primary Road

S1200 Secondary Road

Usage

findRoads(AOI = FALSE)

Arguments

AOI
A Spatial* or simple features geometry, can be piped from getAOI

Value

a list() of minimum length 2: AOI and tiger

Author(s)

Mike Johnson

Examples

## Not run: 
AOI  = getAOI(clip = list("UCSB", 10, 10)) %>% findRoads()

## End(Not run)

[Package HydroData version 0.0.7 Index]