findSSURGO {HydroData}	R Documentation

Find SSURGO Maps and Mapunit Aggregated Attributes

Description

SSURGO is a geospatial database of soils produced by the Natural Resources Conservation Service (NRCS). findSSURGO returns a SpatialPolygonsDataFrame of SSURGO map units and the Mapunit aggregated attributes (muaggatt table) associated with each MUKEY. In total 43 attributes are returned, all of which can be found in the [SSURGO documentation](https://sdmdataaccess.sc.egov.usda.gov/documents/TableColumnDescriptionsReport.pdf). The MUAGGATT values cover a wide range of soil realted questions however if a user is interested in the full range of tabular data associated with the SSURGO map units we suggest checking out the FedData package.

Usage

findSSURGO(AOI)

Arguments

AOI
A Spatial* or simple features geometry, can be piped from getAOI

Value

a list() of minimum length 2: AOI and ssurgo

Author(s)

Mike Johnson

Examples

## Not run: 
AOI  = getAOI(clip = list("UCSB", 10, 10)) %>% findSSURGO()

## End(Not run)

[Package HydroData version 0.0.7 Index]