Find Meterologic Stations from the U.S. Climate Reference Network

Query station information from U.S. Climate Reference Network ([USCRN](https://www.ncdc.noaa.gov/crn/)) and the U.S. Regional Climate Reference Network ([USRCRN](https://www.ncdc.noaa.gov/crn/usrcrn/)). This networks has been established to monitor present and future climatic trends and variability to better define climatic change over the United States. The rogram estimates a fully developed network of about 250 geographical locations are in place. findUSCRN returns a SpatailPointsDataframe of all stations within an AOI. Data comes from the NOAA and includes the following attributes:

'WBAN' : character Weather Bureau–Army–Navy (WBAN) identifiers

'COUNTRY' : character Country FIPS code

'STATE' : integer USA state

'LOCATION' : integer Location description

'VECTOR' : character Location descriptive vecotr

'NAME' : integer Station names

'ELEVATION' : integer Elavation

'STATUS' : integer Is the station currently commissioned

'COMMISSIONING' : integer Date of station the station commissioning

'CLOSING' : character Date of station closing (blank = still running)

'OPERATION' : integer Is station operational?

'PAIRING' : integer Is the station paired with another networks

'NETWORK' : integer USCRN or USRCRN station

findUSCRN(AOI, ids = FALSE)

Arguments

AOI
A Spatial* or simple features geometry, can be piped from getAOI

ids
If TRUE, a vector of unique acis station IDs is added to retuned list (default = FALSE)

Value

a list() of minimum length 2: AOI and and acis

Examples

# NOT RUN {
sta = getAOI(state = "CO", county = "El Paso") %>% findUSCRN()
# }