findWaterPoly {HydroData}	R Documentation

Find Water Polygon

Description

A function to join all NHD water featrues into a single polygon. Flowlines are sized according to there Strahler stream order multiplied by 10 m. Useful for viualization and masking operations where a water surface is needed.

Usage

findWaterPoly(AOI)

Arguments

AOI
an AOI

Value

a single polygon of all NHD water features.

Author(s)

Mike Johnson

Examples

## Not run: 
getAOI(state = 'CO', county = 'El Paso') %>% findWaterPoly()

## End(Not run)

[Package HydroData version 0.0.7 Index]