findWaterbodies {HydroData}	R Documentation

Find National Hydrography Dataset Waterbodies

Description

findWaterbodies returns a SpatialPolgonsDataframe of all NHDwaterbodies within an AOI. Data comes from the USGS CIDA server and contain 23 attributes, perhaps most notably:

'objectid' : integer Integer value that uniquely identifies the waterbody of each feature in the NHD

'comid' : character The COMID draining into the feature

'fdate': POSITct Date of last feature modification

'gnis_id' : character Unique identifier assigned by GNIS

'gnis_name' : character Proper name, term, or expression by which a particular geographic entity is known

'meandepth' : numeric Mean depth of the waterbody

'lakevolume' : numeric Total waterbody volume

'maxdepth' : character Maximum depth of waterbody

'meanused' : integer The average amount of water used

Usage

findWaterbodies(AOI = NULL, area = NULL)

Arguments

AOI
A Spatial* or simple features geometry, can be piped from getAOI

ids
If TRUE, a vector of waterbody IDs is added to retuned list (default = FALSE)

Value

a list() of minimum length 2: AOI and waterboies

Author(s)

Mike Johnson

Examples

## Not run: 
getAOI(clip = "Tuscaloosa") %>% findWaterbodies()

## End(Not run)

[Package HydroData version 0.0.7 Index]